% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_variable_to_data_set.R
\name{add_iso3_to_data_set}
\alias{add_iso3_to_data_set}
\title{Adds iso3 to a given data set}
\usage{
add_iso3_to_data_set(x)
}
\arguments{
\item{x}{A data.table containing a column called "location_code".}

\item{location_reference}{A location reference data.table.}
}
\value{
A data.table containing an extra column called "iso3".
}
\description{
Adds iso3 to a given data set
}
\examples{
data <- data.table(location_code = c("norge", "county03", "blah"))
spldata::add_iso3_to_data_set(data)
print(data)
}
