% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\docType{data}
\name{config}
\alias{config}
\title{Environment containing configuration variables}
\format{
An object of class \code{environment} of length 1.
}
\usage{
config
}
\description{
Environment containing configuration variables
}
\examples{
print(ls(spldata::config))
for(i in names(spldata::config)){
  print(spldata::config[[i]])
}
}
\keyword{datasets}
