% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{denmark_locations_names_b2020}
\alias{denmark_locations_names_b2020}
\title{Names of areas in Denmark that existed in 2020.}
\format{
\describe{
\item{iso3}{ISO3 code.}
\item{granularity_geo}{Nation (NUTS1), region (NUTS2)}
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/NUTS_statistical_regions_of_Denmark}
}
\usage{
denmark_locations_names_b2020
}
\description{
Names of areas in Denmark that existed in 2020.
}
\keyword{datasets}
