% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{denmark_population_by_age_b2020}
\alias{denmark_population_by_age_b2020}
\title{Population in Denmark}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 16 rows and 5 columns.
}
\usage{
denmark_population_by_age_b2020
}
\description{
Population in Denmark
}
\keyword{datasets}
