% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_nordic.R
\docType{data}
\name{finland_locations_names_b2020}
\alias{finland_locations_names_b2020}
\title{Names of areas in Finland that existed in 2020.}
\format{
\describe{
\item{iso3}{ISO3 code.}
\item{granularity_geo}{nation, hospital district}
\item{location_code}{Location code (hospital district.}
\item{location_name}{Location name (hospital district.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/NUTS_statistical_regions_of_Finland}
}
\usage{
finland_locations_names_b2020
}
\description{
Names of areas in Finland that existed in 2020.
}
\keyword{datasets}
