% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{finland_population_by_age_b2020}
\alias{finland_population_by_age_b2020}
\title{Population in Finland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 44 rows and 5 columns.
}
\usage{
finland_population_by_age_b2020
}
\description{
Population in Finland
}
\keyword{datasets}
