% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_get_granularity_geo.R
\name{get_granularity_geo}
\alias{get_granularity_geo}
\title{get_granularity_geo}
\usage{
get_granularity_geo(x, location_reference = NULL)
}
\arguments{
\item{x}{Datatable}

\item{location_reference}{A location reference data.table}
}
\description{
get_granularity_geo
}
