% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_nordic.R
\docType{data}
\name{iceland_locations_names_b2020}
\alias{iceland_locations_names_b2020}
\title{Names of areas in Iceland that existed in 2020.}
\format{
\describe{
\item{iso3}{Location code.}
\item{granularity_geo}{nation}
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/NUTS_statistical_regions_of_Iceland}
}
\usage{
iceland_locations_names_b2020
}
\description{
Names of areas in Iceland that existed in 2020.
}
\keyword{datasets}
