% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{iceland_population_by_age_b2020}
\alias{iceland_population_by_age_b2020}
\title{Population in Iceland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2 rows and 5 columns.
}
\usage{
iceland_population_by_age_b2020
}
\description{
Population in Iceland
}
\keyword{datasets}
