% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icpc2_codes.R
\docType{data}
\name{icpc2_codes}
\alias{icpc2_codes}
\title{ICPC-2: The International Classification of Primary Care (Norwegian, English)}
\format{
\describe{
\item{Code}{ICPC2 code}
\item{title_nb}{Description of the code in Norwegian.}
\item{title_en}{Description of the code in English.}
}
}
\source{
\url{https://ehelse.no/kodeverk/icpc-2.den-internasjonale-klassifikasjonen-for-primaerhelsetjenesten}
}
\usage{
icpc2_codes
}
\description{
Norwegian data is from 2020, English data is from 2018. (Link to be included)
}
\keyword{datasets}
