% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{location_code_to_granularity_geo}
\alias{location_code_to_granularity_geo}
\title{location_code_to_granularity_geo}
\usage{
location_code_to_granularity_geo(x, location_reference = NULL)
}
\arguments{
\item{x}{Datatable}

\item{location_reference}{A location reference data.table}
}
\description{
location_code_to_granularity_geo
}
