% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_code_to.R
\name{location_code_to_iso3}
\alias{location_code_to_iso3}
\title{Convert location_code to iso3}
\usage{
location_code_to_iso3(x)
}
\arguments{
\item{x}{datatable}
}
\value{
Character vector the same length as x, containing the corresponding iso3
}
\description{
Convert location_code to iso3
}
\examples{
spldata::location_code_to_iso3(c("norge", "county03"))
}
