% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\docType{data}
\name{nb}
\alias{nb}
\title{Norwegian characters in unicode}
\format{
An object of class \code{list} of length 6.
}
\usage{
nb
}
\description{
Norwegian characters in unicode
}
\examples{
print(spldata::nb)
}
\keyword{datasets}
