% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_locations_hierarchy.R
\name{nor_locations_hierarchy_from_to}
\alias{nor_locations_hierarchy_from_to}
\title{Location hierarchies in Norway}
\usage{
nor_locations_hierarchy_from_to(
  from,
  to,
  include_to_name = FALSE,
  border = spldata::config$border_nor
)
}
\arguments{
\item{from}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{to}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{include_to_name}{Do you want to include the name of the 'to' location?}

\item{border}{The year in which Norwegian geographical boundaries were designated.}
}
\value{
Data.table containing the columns:
- from_code
- to_code
- to_name (if include_to_name==TRUE)
}
\description{
Calculates the relationship between different locations in Norway, according
to geographic granularity. For example, which municipalities are inside which counties.
}
\examples{
spldata::nor_locations_hierarchy_from_to(from="wardoslo", to="county")
spldata::nor_locations_hierarchy_from_to(from="municip", to="baregion")
}
