% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_locations_redistricting.R
\name{nor_locations_redistricting}
\alias{nor_locations_redistricting}
\title{All redistricting in Norway}
\usage{
nor_locations_redistricting(border = spldata::config$border)
}
\arguments{
\item{border}{The year in which Norwegian geographical boundaries were designated.}
}
\value{
\format{data.table}
\describe{
\item{location_code_current}{The location code per today.}
\item{location_code_original}{The location code as of 'year'.}
\item{calyear}{The year corresponding to 'county_code_original'.}
\item{weighting}{The weighting that needs to be applied.}
\item{granularity_geo}{nation/county/municip/wardbergen/wardoslo/wardstavanger/wardtrondheim/missingwardbergen/missingwardoslo/missingwardstavanger/missingwardtrondheim/notmainlandcounty/notmainlandmunicip/missingcounty}
}
}
\description{
This function returns a dataset that is used to transfer "original" datasets
to the 2020 borders.
}
\examples{
spldata::nor_locations_redistricting()
}
