% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_loc_hierarchy.R
\docType{data}
\name{norway_locations_hierarchy_b2020}
\alias{norway_locations_hierarchy_b2020}
\title{Hierarchy of different levels in Norway (2020 borders).
#'
Last updated 2021-02-15}
\format{
\describe{
\item{municip_code}{The location code per today.}
\item{municip_name}{The location code as of 'year'.}
\item{baregion_code}{The location code per today.}
\item{baregion_name}{The location code as of 'year'.}
\item{county_code}{The location code as of 'year'.}
\item{county_name}{The location code as of 'year'.}
\item{region_code}{The location code as of 'year'.}
\item{region_name}{The location code as of 'year'.}
\item{faregion_name}{The location code as of 'year'.}
\item{faregion_code}{The location code as of 'year'.}
\item{ward_code}{The location code as of 'year'.}
\item{ward_name}{The location code as of 'year'.}
\item{missingward_code}{The location code as of 'year'.}
\item{missingward_name}{The location code as of 'year'.}
\item{notmainlandmunicip_code}{The location code as of 'year'.}
\item{notmainlandmunicip_name}{The location code as of 'year'.}
\item{notmainlandcounty_code}{The location code as of 'year'.}
\item{notmainlandcounty_name}{The location code as of 'year'.}
\item{missingmunicip_code}{The location code as of 'year'.}
\item{missingmunicip_name}{The location code as of 'year'.}
\item{missingcounty_code}{The location code as of 'year'.}
\item{missingcounty_name}{The location code as of 'year'.}
}
}
\usage{
norway_locations_hierarchy_b2020
}
\description{
Hierarchy of different levels in Norway (2020 borders).
#'
Last updated 2021-02-15
}
\keyword{datasets}
