% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_loc_hierarchy.R
\name{norway_locations_hierarchy_from_to}
\alias{norway_locations_hierarchy_from_to}
\title{Hierarchies in Norway (programmable borders).}
\usage{
norway_locations_hierarchy_from_to(
  from,
  to,
  include_to_name = FALSE,
  border = spldata::config$border
)
}
\arguments{
\item{from}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{to}{wardoslo, wardbergen, wardtrondheim, wardstavanger, municip, baregion, county, region, faregion, notmainlandmunicip, notmainlandcounty, missingmunicip, missingcounty}

\item{include_to_name}{Do you want to include the name of the 'to' location?}

\item{border}{The border year}
}
\description{
Hierarchies in Norway (programmable borders).
}
\examples{
norway_locations_hierarchy_from_to(from="wardoslo", to="county")
norway_locations_hierarchy_from_to(from="municip", to="baregion")
}
