% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_loc_redistricting.R
\name{norway_locations_redistricting}
\alias{norway_locations_redistricting}
\title{All redistricting in Norway (programable borders).}
\usage{
norway_locations_redistricting(
  include_year = TRUE,
  border = spldata::config$border
)
}
\arguments{
\item{include_year}{Do you want to include redistricting by year?}

\item{border}{The border year}
}
\description{
All redistricting in Norway (programable borders).
}
\examples{
norway_locations_redistricting()
}
