% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_loc_redistricting.R
\docType{data}
\name{norway_locations_redistricting_b2020}
\alias{norway_locations_redistricting_b2020}
\title{Redistricting in Norway (2020 borders).}
\format{
\describe{
\item{location_code_current}{The location code per today.}
\item{location_code_original}{The location code as of 'year'.}
\item{calyear}{The year corresponding to 'county_code_original'.}
\item{weighting}{The weighting that needs to be applied.}
\item{granularity_geo}{nation/county/municip/wardbergen/wardoslo/wardstavanger/wardtrondheim/missingwardbergen/missingwardoslo/missingwardstavanger/missingwardtrondheim/notmainlandcounty/notmainlandmunicip/missingcounty}
}
}
\usage{
norway_locations_redistricting_b2020
}
\description{
This dataset is used to transfer "original" datasets to the 2020 borders.
}
\details{
Last updated 2021-11-15
}
\keyword{datasets}
