% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{data}
\name{norway_population_by_age_b2020}
\alias{norway_population_by_age_b2020}
\title{Population in Norway (2020 borders).}
\format{
\describe{
\item{year}{Year.}
\item{location_code}{The location code.}
\item{granularity_geo}{National/County/Municipality/BAregion.}
\item{age}{1 year ages from 0 to 105.}
\item{pop_jan1}{Number of people as of 1st of January.}
\item{imputed}{FALSE if real data. TRUE if it is the last real data point carried forward.}
}
}
\source{
\url{https://www.ssb.no/en/statbank/table/07459/tableViewLayout1/}
}
\usage{
norway_population_by_age_b2020
}
\description{
We conveniently package population data taken from Statistics Norway.
This data is licensed under the Norwegian License for
Open Government Data (NLOD) 2.0.
}
\details{
This dataset contains national/county/municipality/ward (city district) level population data
for every age (0 to 105 years old). The national level data is from year 1846, while all the
other levels have data from 2005.

The counties and municipalities are updated for the 2020 borders.
}
\keyword{datasets}
