% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\docType{data}
\name{se}
\alias{se}
\title{Swedish characters in unicode}
\format{
An object of class \code{list} of length 4.
}
\usage{
se
}
\description{
Swedish characters in unicode
}
\keyword{datasets}
