% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_environments.r
\name{set_config}
\alias{set_config}
\title{set_config}
\usage{
set_config(border = 2020, use_current_year_as_1900_pop = FALSE)
}
\arguments{
\item{border}{The year}

\item{use_current_year_as_1900_pop}{Replaces the year 1900's population data with the current year's population data}
}
\description{
set_config
}
