% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scr_nordic.R
\docType{data}
\name{sweden_locations_names_b2020}
\alias{sweden_locations_names_b2020}
\title{Names of areas in Sweden that existed in 2020.}
\format{
\describe{
\item{iso3}{ISO3 code.}
\item{granularity_geo}{Nation (NUTS1), county (NUTS3)}
\item{location_code}{Location code.}
\item{location_name}{Location name.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/NUTS_statistical_regions_of_Sweden}
}
\usage{
sweden_locations_names_b2020
}
\description{
Names of areas in Sweden that existed in 2020.
}
\keyword{datasets}
