% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nordic.R
\docType{data}
\name{sweden_population_by_age_b2020}
\alias{sweden_population_by_age_b2020}
\title{Population in Sweden}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 44 rows and 5 columns.
}
\usage{
sweden_population_by_age_b2020
}
\description{
Population in Sweden
}
\keyword{datasets}
