% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\docType{data}
\name{config}
\alias{config}
\title{An environment containing configuration variables}
\format{
An object of class \code{environment} of length 1.
}
\usage{
config
}
\description{
Available configuration variables:
\itemize{
\item border_nor (default 2020): The year in which Norwegian geographical boundaries were designated.
}
}
\examples{
print(ls(spldata::config))
for(i in names(spldata::config)){
  cat(i, ":", spldata::config[[i]], "\n")
}
}
\keyword{datasets}
