% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_code_to.R
\name{location_code_to_granularity_geo}
\alias{location_code_to_granularity_geo}
\title{Convert location_code to granularity_geo}
\usage{
location_code_to_granularity_geo(x, location_reference = NULL)
}
\arguments{
\item{x}{Either a vector, or a data.frame/data.table containing a column called "location_code"}

\item{location_reference}{A location reference data.table}
}
\value{
Character vector the same length as x, containing the corresponding granularity_geo
}
\description{
Convert location_code to granularity_geo
}
\examples{
spldata::location_code_to_granularity_geo(c("norge", "county03"))
}
