% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nor_population_by_age_cats.R
\name{nor_population_by_age_cats}
\alias{nor_population_by_age_cats}
\title{Population in Norway by categories}
\usage{
nor_population_by_age_cats(
  cats = NULL,
  include_total = TRUE,
  border = spldata::config$border
)
}
\arguments{
\item{cats}{A list containing vectors that you want to categorize}

\item{include_total}{Boolean. Should 'total' be included as an age cat?}

\item{border}{The border year}
}
\value{
A data.table containing
}
\description{
A function that easily categorizes the populations for you.
}
\examples{
nor_population_by_age_cats(cats = list(c(1:10), c(11:20)))
nor_population_by_age_cats(cats = list("one to ten" = c(1:10), "eleven to twenty" = c(11:20)))
nor_population_by_age_cats(cats = list(c(1:10), c(11:20), "021p"=c(21:200)))
}
