% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_aa_environments.r
\name{set_config}
\alias{set_config}
\title{Set options in the package config}
\usage{
set_config(border_nor = NULL)
}
\arguments{
\item{border_nor}{The year in which Norwegian geographical boundaries were designated.}
}
\value{
Nothing. Side effect of setting the \code{config} environment.
}
\description{
Set options in the package config
}
